Added sub-chapter "Tweaking the INI files".
I have been away from the guide for over 4 months. It's been a bit difficult to bring myself up to speed again, but it feels very, very good.

This chapter is a bit strange, because I had not planned for it properly, but I think it's fairly ok (I hope so, at least).