Added sub-chapter "Setting up the foundation for our modding".
This chapter has been recovered from the vol. 1... But it has nothing to do with the former. Mod Organizer has a very different way of doing things.

I have relocated the chapter just before the real modding (cleaning the master files). I felt that's the right place, because FOSE it's the foundation for modding, after all.