The first part of the "Bashing all together" sub-chapter is complete ("Generating your own bashed patch"), and it has been a fucking difficult one. Creating bashed patches is not for the faint of heart.
I have yet to complete the second part ("Using a pre-generated bashed patch"), which is going to be a lot easier. The good thing is I only need to write these detailed instructions once, because in the rest of the major DEFCON chapters I am going to use the pre-generated versions of all my merged and bashed patches.

Ah, I added some metadata to the merged patch (author and description), so it shows up when loaded in Wrye Flash.