The chapter about DEFCON 01 patches has undergone a major overhaul.
I didn't feel comfortable with the structure I gave to this chapter. I think it was confusing and too large. For starters, why talking about creating your own merged/bashed patch if this is the less desirable option? I am creating the tweaked patches for a reason.

So, I have moved the creation process of the patches to separate appendices, one for merged and one for bashed patches. The sideloading of the provided merged patch remains in place, at the end of the DEFCON 01 chapter, although it is now part of a level 3 chapter which includes both patches.

There is some work to do in the appendices, because I would like to rewrite them later on, when there is some weight in the mod stack. However, the current images serve as placeholders, for now.

I need to write now the sub-chapter about sideloading the provided bashed patch, but it's going to be easy, using what I already have about the merged patch.