Added sub-chapter "Chapter recap".
I have added a small chapter with a quick recap of the DEFCON 01 major chapter. This chapter was not planned, but I thought it would be nice to have an informative chapter which linked the DEFCON 01 and DEFCON 02 chapters, even if you decide to stop modding the game at this point.