Added blocks with stacking and load order at the end of each modding chapter.
I didn't feel comfortable with the links to the load order scattered through the text. They weren't easy to spot, and at a certain point, I needed to add _both_ stacking order and load order links.

I have cleared up the mess adding a block at the end of each chapter which modifies the stacking order or the load order, with links to the corresponding lists, so you can keep track of the contents of the left and right pane in MO.