Added sub-chapter "Face-washing the HUD".
The modding process we have performed has not been very impressive because all the things we have done work under the hood. In this chapter we are starting to see real differences in the look of the game. The user interface has changed dramatically, improving the user experience.

This has been a long chapter, by the way, but I like the way it looks. I think it's solid and thorough.