Fixed the title in images with commas.
I wrote the image titles withous surrounding double-quotes, so
Asciidoctor was considering that the text right to the comma was, in
fact, the width of the image.

Consider this snippet:

image::foo.png[Foo,10,20]

That means "load foo.png with the title 'Foo', 10 pixels width and 20
pixels height".

That's sensible. What's not sensible is doing this shit:

image::foo.png[Foo, truly]

See? Fortunately, the fix is a real trifle:

image::foo.png["Foo, truly"]