Finally! Weg get rid of the old compilation system in favor of the full-fledged Gradle build system.
The build script performs the following tasks:

1. Compiles the SASS stylesheet template into its CSS equivalent.
2. Creates the HTML5 and PDF backends of the book.
3. Copies back the CSS stylesheet into the build folder.

Sadly, I have not been able to use the Compass/SASS Gradle plugin due to
an internal error, so I have to use the Asciidoctor stylesheet factory
compilation method.

On top of that, the current version of the stylesheet factory does not
work, so I have had to use an old version which I have embedded into the
project folder, adding it to the source control.