Added sub-chapter "Everything depends on the point of view".
It's been a long time without a new sub-chapter, and I just realized
that this book has been two years in the making.

Two fucking years.

Well, there's a lot of work to do, so let's get rid of distractions.

This chapter has been an easy one (or it should have been, provided that
I had had my muscle memory well honed). It's just about two fixes in 1st
and 3rd person view. No big deal, but it's good to be back in the game.