Added sub-chapter "Cleaning up your faithful companion".
This chapter was in the vol. 1, although I've chosen to write it from
scratch. Nothing fancy here, except the Pip-Boy cleaner, which has a lot
of possibilities. In the vol. 1 I remember using several of them
(removing vintage artifacts, for example), but in this version I'll go
with just the map tint removal. I've come to love the retro look.